refactor: Operaorを別モジュールに分離
